// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/productivityDashboardModel.ts
export class ProductivityDashboardModel {
    pageSize: number;
    pageNumber: number;
    searchText: string;
    sortBy: string;
    sortDirectionAsc: boolean;
    selectedDate: string;
    type: string;
    projectId: string;
    ownerUserId: string;
    userStoryStatusId: string;
    deadLineDateFrom: string;
    deadLineDateTo: string;
    isUserstoryOutsideQa: boolean;
    entityId: string;
    indexType: string;
    isReportingOnly: boolean;
    isAll: boolean;
    isMyself: boolean;
    isFromDrilldown: boolean;
    columns: any;
    state: any;
    dateFrom: any;
    dateTo: any;
} 